+++ src/main/java/com/invadermonky/overburdened/compat/InitCompat.java
@@ -5,6 +5,7 @@ import com.invadermonky.overburdened.api.custom.ShulkerBoxCapabilityHandler;
 import com.invadermonky.overburdened.compat.actuallyadditions.ItemBagCapabilityHandler;
 import com.invadermonky.overburdened.compat.hbm.HBMCapabilityHandler;
 import com.invadermonky.overburdened.compat.immersiveengineering.IECapabilityHandler;
+import com.invadermonky.overburdened.compat.ironchests.IronChestCapabilityHandler;
 import com.invadermonky.overburdened.compat.travelersbackpack.TravelersBackpackCapabilityHandler;
 import com.invadermonky.overburdened.utils.libs.ModIds;
 
@@ -18,6 +19,8 @@ public class InitCompat {
             OverburdenedAPI.registerCustomCapabilityHandler(new HBMCapabilityHandler());
         if (ModIds.immersive_engineering.isLoaded)
             OverburdenedAPI.registerCustomCapabilityHandler(new IECapabilityHandler());
+        if (ModIds.iron_chests.isLoaded)
+            OverburdenedAPI.registerCustomCapabilityHandler(new IronChestCapabilityHandler());
         if (ModIds.travelers_backpack.isLoaded)
             OverburdenedAPI.registerCustomCapabilityHandler(new TravelersBackpackCapabilityHandler());
     }